Fetch maxes in dashboard, since one rep max might not be rendered before workout program
