+++ src/App.tsx
@@ -1,6 +1,7 @@
 import { Route, Routes, Navigate } from "react-router-dom";
 import { useAuthUpdate } from "./contexts/AuthContext";
 import { ModalProvider } from "./contexts/ModalContext";
+import { UserMaxesProvider } from "./contexts/UserMaxesContext";
 import { ToastContainer } from "react-toastify";
 import "react-toastify/dist/ReactToastify.min.css";
 import HomePage from "./scenes/home/HomePage";
@@ -17,30 +18,32 @@ const App = () => {
 
   return (
     <>
-      <ModalProvider>
-        <Nav />
-        <Routes>
-          <Route path="/" element={<HomePage />} />
-          <Route
-            path="/login"
-            element={
-              !isAuthenticated ? (
-                <Login setAuth={setAuth} />
-              ) : (
-                <Navigate to="/dashboard/workout_program" />
-              )
-            }
-          />
-          <Route
-            path="/dashboard/*"
-            element={
-              isAuthenticated ? <DashboardPage /> : <Navigate to="/login" />
-            }
-          />
-        </Routes>
-        <Footer />
-        <ToastContainer />
-      </ModalProvider>
+      <UserMaxesProvider>
+        <ModalProvider>
+          <Nav />
+          <Routes>
+            <Route path="/" element={<HomePage />} />
+            <Route
+              path="/login"
+              element={
+                !isAuthenticated ? (
+                  <Login setAuth={setAuth} />
+                ) : (
+                  <Navigate to="/dashboard/workout_program" />
+                )
+              }
+            />
+            <Route
+              path="/dashboard/*"
+              element={
+                isAuthenticated ? <DashboardPage /> : <Navigate to="/login" />
+              }
+            />
+          </Routes>
+          <Footer />
+          <ToastContainer />
+        </ModalProvider>
+      </UserMaxesProvider>
     </>
   );
 };
